'hello'

+++ YellowstonePathology/UI/AORemover.cs
@@ -16,10 +16,14 @@ namespace YellowstonePathology.UI
             YellowstonePathology.Business.Rules.MethodResult methodResult = new Business.Rules.MethodResult();
             accessionOrder.AccessionLock.ReleaseLock();
             
-            YellowstonePathology.Business.ClientOrder.Model.ClientOrder clientOrder = Business.Persistence.DocumentGateway.Instance.PullClientOrder(accessionOrder.ClientOrderId, writer);
-            YellowstonePathology.Business.Persistence.DocumentGateway.Instance.DeleteDocument(clientOrder, writer);
+            if(string.IsNullOrEmpty(accessionOrder.ClientOrderId) == false)
+            {
+                YellowstonePathology.Business.ClientOrder.Model.ClientOrder clientOrder = Business.Persistence.DocumentGateway.Instance.PullClientOrder(accessionOrder.ClientOrderId, writer);
+                YellowstonePathology.Business.Persistence.DocumentGateway.Instance.DeleteDocument(clientOrder, writer);                
+            }
+
             YellowstonePathology.Business.Persistence.DocumentGateway.Instance.DeleteDocument(accessionOrder, writer);
-            
+
             return methodResult;
         }
